+++ src/utils/mod.rs
@@ -7,9 +7,7 @@ mod json_object;
 mod plurals;
 mod profile;
 
-pub(crate) use app_url::{
-    app_project_url, app_project_url_with_encoded_path, app_project_url_with_query,
-};
+pub(crate) use app_url::{app_project_url, app_project_url_with_encoded_path};
 pub use duration::parse_duration_to_seconds;
 pub use fs_atomic::{write_bytes_atomic, write_text_atomic};
 pub use git::GitRepo;
